+++ newspack-event-logger-nodes.php
@@ -31,6 +31,7 @@ require_once NEWSPACK_EVENT_LOGGER_NODES_DIR . 'vendor/autoload.php';
 $_newspack_event_logger_nodes_load = static function (): void {
 	if ( \defined( 'WP_CLI' ) && \WP_CLI ) {
 		\WP_CLI::add_command( 'nodes reqgrep', '\\Newspack_Event_Logger_Nodes\\CLI\\Reqgrep_Command' );
+		\WP_CLI::add_command( 'nodes ruleset-bench', '\\Newspack_Event_Logger_Nodes\\CLI\\Ruleset_Bench_Command' );
 	}
 
 	\add_action(